ml-service/app/requirements.txt
fastapi
uvicorn[standard]
ultralytics
numpy
pillow
python-multipart
